Docs: Add list of "Guides", and "Support" section to Getting Started
Promote a curated list of guides on the Getting Started page.
his replaces the last of the lost pathways that the Cookbook
ffered on the original website.
https://github.com/qunitjs/qunitjs.com/issues/151
https://web.archive.org/web/20160303175442/http://qunitjs.com/cookbook/#efficient-development
https://github.com/qunitjs/qunitjs.com/blob/v2.10.2/pages/cookbook.html

Improve visibility of the chat room and issue tracker as well, by
pointing to them directly from the Getting Started page. They are
also at the bottom of the homepage, but the main CTA points away from
there so this seems worth repeating in prose here.