Core: Fix infinite loop crash when mixing test.only() with module.only()
I noticed that the first branch of `addOnlyTest()` was uncovered in
test coverage reports. In expanding the `only-module-then-test.js`
fixture to cover this (by adding a late test.only call), which
was originally added in https://github.com/qunitjs/qunit/pull/1658.

I decided to also add an early test.only call to see what would happen.
The outcome is irrelevant/ambigious, but it seemed worth observing.
To my surprise, it created an infinite loop.

This further convinces me that we should make "early errors" show
up in the HTML Reporter (TODO added in 37f6e4b6613), and subsequently
get rid of this re-entry hack. The TapReporter and QUnit CLI are
already able to render early errors. The HtmlReporter not yet. It's
not generally a priority imho, but, because "No tests" is a common
scenario, if we utilize an early error to report it, that makes it
a priority to support in HtmlReporter as otherwise there'd be no
visible feedback for it. It's okay to have to pop open devtools when
causing an early syntax error or something like that in your own code,
but when using the UI to match no tests, there should at least be
some kind of visible feedback and not a blank page.