Test: Convert /test/cli/TapReporter.js to /test/main
Since commit f8948c96fdcef0b, the TapReporter is exposed publicly
as QUnit.reporters.tap and it is used in browser automations,
and for SpiderMonkey environments.

Let's make sure this not only passes synax and other compat checks
in eslint and rollup for the main main build, but is also covered in
the main test suite.