HTML Reporter: Faster "Hide passed" toggling on large test suites
Optimize the hidepassed click handler for large test suites by
avoiding internal function call overhead for the iterator, as well
as copying cost for the array. Shaves off a few milliseconds on the
q4000 demo overall from 10ms to 5ms (Chrome), and seems to move some
unattributed costs to outside the critical path.