Docs: Improve browser runner docs
Move "Test results explained" and failure example higher up on the
page.

Mention "devtools".

Change failure example to include two passings to be more realistic.

Change failure example to have 3 passing assertions.

Previously:
> - test name (1,2,3)
>   1. a
>   2. [fail] b
>   3. c

 Now:
> - test name (1,3,4)
>   1. a
>   2. [fail] b
>   3. c
>   4. d

The "1,2,3" count was imho non-obvious in meaning. While the meaning
is explained in prose, with 4 assertions it is much more intuitive
to identify that there is 1 failure, 3 passing, and 4 in total.