Docs: Demonstrate how to load custom reporter from local script
It is not currently valid to pass `./foo.js` to `--reporter` because
cli/run.js and cli/find-reporter.js would be running require() relative
to where the qunit package is installed instead of the project directory.

We could make this work in the future, but this is how it works today.