Core: Make initial "unnamed" QUnit.config.currentModule complete
Before this change, the unnamed module object was not created by
`createModule()`. Instead, it was declared in config.js as the first
value for `config.currentModule`, with only some of the properties.

It had enough to satisfy internal execution requirements, but some
publicly observable stuff like `moduleId` and `testEnvironment` were
not defined.

Misc changes and optimisations:

* `suiteReport` is now injected instead of overridden afterwards.

* `moduleStack` state is moved to config along with other internal
  state. This removes one of the last remaining state that wasn't
  resettable.

* When creating child modulese, avoid allocating unused `env = {}`.

* When creating module objects, avoid needless copying of the
  moduleStack via Array slice().

* In test.js, simplify how we fill `completedModules` by using the
  same loop for both parenet modules and the current module alike,
  thus removing duplicate code.