Core: Always define globalThis.QUnit
== Background

In QUnit 2.x, we always set the QUnit global in browser contexts, but
outside browsers, we only set the global if CJS/AMD wasn't detected.
Then, in the QUnit CLI, we set the global anyway, and we expect other
test runners to do the same, since that's how most QUnit tests are
written, including for Node.js targets.

In practice, the QUnit global would only be missing in custom test
runners that 1) target Node.js, 2) require/import qunit.js directly,
and 3) don't export it as a global.

I'm aware many communities import 'qunit' directly in each test file
for improved type support. That's great, and avoids needing to rely
on globals. But, that's not a requirement I want to impose on everyone,
especially for simple no-build-step and browser-facing projects.

== Why

* Improve portability between test runners.
  Remove the last edge case where the QUnit global can be undefined.
  Make it QUnit's responsiblity to define this reliably, as indeed it
  almost always already does. Remove this as undocumented requirement
  for specific test runners to patch up on their end.

  In light of Karma deprecation, and emergence of more general purpose
  TAP runners, I'd like to remove this as factor that might make/break
  QUnit support in one of those.