HTML Reporter: Revert "Clip qunit-userAgent to one line" (unreleased)
Follows-up d8c2a3ac18. This felt like a hacky compromise and doesn't
quite feel right either way. Let's leave flowing as before for now.

What I did add is to expose the version and user agent as data
attributes, to empower other themes and plugins to play with this
and perhaps come up with a better design.