Core: Restore `QUnit.extend` without deprecation
We need the function internally either way, for its internal features,
such as `undefOnly`.

Too widely used in the ecosystem. Not worth the effort for devs to
have change their code as part of upgrading QUnit, and possibly
stagnating on an older QUnit version over (if not in their own code,
but e.g. a plug-in).

See also:
* 82c4790d07: Remove `extend`.
* 335b040a48: Document removal of `extend`.
* eaa0dbebcb: Restore `assert.push` without deprecation.