Core: Restore IE9-10 support
The cost of supporting IE9-10, given we support IE11 is tiny it is
only a handful of lines of code. In order to reduce the size of the
3.0 release and get it out the door, bring this back.

This also allows it to be used by the latest jQuery 3.x which supports
IE9 still.

Note that an ESM bundle was introduced in QUnit 3.0.0-alpha.4,
which is native ES6 and drops support for IE 9, IE 10 and IE 11, and
indeed all down-compilation for ES5-era browsers. You can use that
in order to get the same benefit in terms of bundle size.

Having said that, QUnit has one of the leanest bundle sizes of any
test framework even with IE9 support included.