Docs: Create a "Community" menu and add "Support & Chat" link
* Rename the About menu to "Community".
* Move "Browser Support" to Guides.
* Write "Support" and "Chat" sections.
* Add "Support & Chat" link to the Community menu.
* Enable Bluesky link in theme footer.

Also sort out the mess that's become of the app.element.io client
which for unknown reasons no longer allows guest viewing, it just
throws up a login wall instead. As workaround, use the app.gitter.im
link, which hosts the same Element Web client, but does still permit
guest viewing.

The reason to keep both is that the gitter.im client is configured
to only allow login via GitHub or GitLab account. This domain is
not registered as universal link for Element native apps, and also
offers no (easy) way to create (or login with) a Matrix.org account.