CLI: Fix TAP compliance for actual/expected indent and skip/todo colors
== Indentation ==

When reporting a failing test, and when the actual/expected data
is an array or object, we just embedded the result of JSON.stringify
as-is without indenting each line correctly. This meant that the
property keys were oddly aligned with the outer object, and that
the closing bracket was actually outside the `---` diagnostic block.

Example:

```
not ok 1 example
  ---
  message: foo
  severitity: failed
  actual: {
  "x": 1,
  "y": 2
}
  expected: {
  "x": 1,
  "y": 3
}
  ...
```

The formatting of the property keys is purely cosmetic and the spec [1]
doesn't care about that. (Although humans do!) But, what the spec
does require is that the diagnostic YAML block ends when there
is a line that doesn't start with two spaces.

As a result, parsing QUnit CLI output with tap-parser [2] did not work
very well when you look closely. It correctly understands the overall
summary and the name and status of each test result. But, the
diagnostic data is lost (interpreted as non-TAP extra lines, equivalent
to random console.log lines).

[1]: https://testanything.org/tap-version-13-specification.html
[2]: https://github.com/tapjs/tapjs/tree/tap-parser%4018.0.0/src/parser

== Colors ==

Reporting of SKIP and TODO directives was also not skip compliant.
When the formatting of "skip" and "todo" was introduced back in 2017,
with https://github.com/qunitjs/js-reporters/pull/95, it made two
mistakes:

```
not ok 1 # TODO hello
ok 2 # SKIP world
```

1. The color code started right before the "#" character,
   like `not ok 1 <cyan># TODO hello</cyan>`.

   This does not follow the spec [1], because a test is only
   considered a todo if "<space>#<space>TODO" appears after the
   (optional) test name, and this isn't the case for us.

   Because of this, todo tests were parsed by tap-parser as genuine
   test failures, thus breaking the end-user result.

2. The test name was printed after "#" instead of before.
   This meant that strictly speaking, the name of the test was the
   empty string "", and the "reason" was set to the test name.

   This is not a big deal. So long as the (optional) reason is
   preserved and presented also by the next TAP tool in chain,
   the user will still be able to see it.