Docs: Move "Plugins" page from Guides menu to Community menu
It was always an awkward fit where it was under Guides, because it
isn't a guide at all, but we didn't have any other menus, and there
wasn't enough space in the design to fit an extra top-evel page.

In d484d145fd84e1 we renamed the "About" menu to "Community", which
makes a great home for the Plugins page!