Test: Move cli-watch.js fixture from cwd to /tmp
Avoid changing files in the current working directory as that means,
if you use `node --watch-path`, it will go into an infinite loop that
always kills the test before it begins.

Also, remove "x.ts" fixture from a generic test because this fails
on Node.js 23 where ".ts" is part of require.extensions by default
even without the ts-node/register loader for TypeScript.

The test failed because we actually do restart by default for changes
in `.ts` files, on Node.js 23, which is great.