Docs: Use parens in arrow functions, and overall reduce arrows funcs
Cut back the use of arrow functins somewhat, especially in migration
guides from older versions, and in generic examples that aren't about
scope.

When we use them, consistency use parentheses around the arguments.

Remove the `hooks => ` and `assert => ` idiom from examples which
have not taken off in practice, and don't jive well with most
practical style guides. It may seem minimal, but it makes the code
not realistic and not directly portable. It also makes it harder to
adopt, e.g. to add an argument such as the case when using
`QUnit.test.each()`.