CLI: Fix TAP compliance for colon in unquoted YAML diag
When an assertion message contained a colon followed by a space,
and otherwise is a single line of ASCII characters, we formatted it
as an unquoted string. This is invalid in YAML 1.2, and thus tap-parser
would ignore the diagnostic block.

This does not affect the TAP result itself (the test name and failure
status are separate from this), and it does not affect the QUnit CLI
output where the information is shown either way.

When passing the CLI output to tap-parser (such as in QTap), it
silently ignored the "diag" block containing the assertion
message/actual/expected value, because it is not valid YAML.

Note that YAML 1.2 does allow literal colons and hash tags in unquoted
strings, when they are not followed by or preceded by a space, but
I'm keeping our version simpler by erring toward quoted strings when
these appear. Just because we could squeeze one more edge case with
plain unquoted strings, doesn't mean we have to. This way keeps our
format a bit more intuitive by making it predictable and easy to
deduce why a string is quoted or not, based solely on the chars used
and not based on obscure YAML internals.

Ref https://yaml.org/spec/1.2.2/#733-plain-style