CLI: Fix TAP compliance for test plan when using `QUnit.test.only()`
We correctly prevent registration, counting, and execution of non-only
tests after the first QUnit.test.only call. And we correctly undo
the execution and registration of any QUnit.test calls before the
first QUnit.test.only call.

But, we did not clear the test *count*. This left the test count
emitted in `runEnd` event, and thus in the "1..X" test plan printed
by the TAP reporter, inaccurate. This has implications for TAP-based
test runners (such as QTap) as those rely on this being accurate to
declare a run as finished, stop waiting for results, and stop the
browser process.